CLN: create core Styler rendering functions and subclass conditional styling applications (#40745)
